+++ src/main/java/com/synopsys/integration/detect/lifecycle/run/operation/OperationFactory.java
@@ -112,8 +112,6 @@ import com.synopsys.integration.detect.workflow.bdio.aggregation.FullAggregateGr
 import com.synopsys.integration.detect.workflow.blackduck.BlackDuckPostOptions;
 import com.synopsys.integration.detect.workflow.blackduck.DetectFontLoader;
 import com.synopsys.integration.detect.workflow.blackduck.bdio.IntelligentPersistentUploadOperation;
-import com.synopsys.integration.detect.workflow.blackduck.bdio.LegacyBdio1UploadOperation;
-import com.synopsys.integration.detect.workflow.blackduck.bdio.LegacyBdio2UploadOperation;
 import com.synopsys.integration.detect.workflow.blackduck.codelocation.CodeLocationWaitCalculator;
 import com.synopsys.integration.detect.workflow.blackduck.codelocation.CodeLocationWaitData;
 import com.synopsys.integration.detect.workflow.blackduck.codelocation.WaitableCodeLocationData;
@@ -358,20 +356,6 @@ public class OperationFactory { //TODO: OperationRunner
         return auditLog.namedInternal("Create Aggregate Options", () -> new AggregateDecisionOperation(detectConfigurationFactory.createAggregateOptions()));
     }
 
-    public final BdioUploadResult uploadBdio1(BlackDuckRunData blackDuckRunData, BdioResult bdioResult) throws OperationException {
-        return auditLog.namedPublic(
-            "Upload Legacy Bdio 1",
-            () -> new LegacyBdio1UploadOperation(blackDuckRunData.getBlackDuckServicesFactory().createBdioUploadService()).uploadBdioFiles(bdioResult)
-        );
-    }
-
-    public final BdioUploadResult uploadBdio2(BlackDuckRunData blackDuckRunData, BdioResult bdioResult) throws OperationException {
-        return auditLog.namedPublic(
-            "Upload Legacy Bdio 2",
-            () -> new LegacyBdio2UploadOperation(blackDuckRunData.getBlackDuckServicesFactory().createBdio2UploadService()).uploadBdioFiles(bdioResult)
-        );
-    }
-
     public final BdioUploadResult uploadBdioIntelligentPersistent(BlackDuckRunData blackDuckRunData, BdioResult bdioResult, Long timeout) throws OperationException {
         return auditLog.namedPublic(
             "Upload Intelligent Persistent Bdio",
@@ -477,14 +461,13 @@ public class OperationFactory { //TODO: OperationRunner
         return auditLog.namedPublic("Create Risk Report File", "RiskReport", () -> {
             DetectFontLoader detectFontLoader = detectFontLoaderFactory.detectFontLoader();
             ReportService reportService = creatReportService(blackDuckRunData);
-            File createdPdf = reportService.createReportPdfFile(
+            return reportService.createReportPdfFile(
                 reportDirectory,
                 projectVersionWrapper.getProjectView(),
                 projectVersionWrapper.getProjectVersionView(),
                 detectFontLoader::loadFont,
                 detectFontLoader::loadBoldFont
             );
-            return createdPdf;
         });
     }
 
@@ -784,7 +767,6 @@ public class OperationFactory { //TODO: OperationRunner
 
     public void createAggregateBdio2File(AggregateCodeLocation aggregateCodeLocation) throws OperationException {
         auditLog.namedInternal("Create Bdio Code Locations", () -> {
-            DetectBdioWriter detectBdioWriter = new DetectBdioWriter(new SimpleBdioFactory(), detectInfo);
             new CreateAggregateBdio2FileOperation(new Bdio2Factory(), detectInfo).writeAggregateBdio2File(aggregateCodeLocation);
         });
     }