+++ src/main/java/com/synopsys/integration/detect/workflow/bdio/CreateBdio1FilesOperation.java
@@ -7,10 +7,14 @@ import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.synopsys.integration.bdio.BdioNodeFactory;
+import com.synopsys.integration.bdio.BdioPropertyHelper;
 import com.synopsys.integration.bdio.SimpleBdioFactory;
 import com.synopsys.integration.bdio.graph.DependencyGraph;
 import com.synopsys.integration.bdio.graph.DependencyGraphUtil;
 import com.synopsys.integration.bdio.graph.ProjectDependencyGraph;
+import com.synopsys.integration.bdio.model.BdioBillOfMaterials;
+import com.synopsys.integration.bdio.model.BdioProject;
 import com.synopsys.integration.bdio.model.SimpleBdioDocument;
 import com.synopsys.integration.bdio.model.dependency.ProjectDependency;
 import com.synopsys.integration.bdio.model.externalid.ExternalId;
@@ -37,13 +41,30 @@ public class CreateBdio1FilesOperation {
         List<UploadTarget> uploadTargets = new ArrayList<>();
         for (BdioCodeLocation bdioCodeLocation : bdioCodeLocationResult.getBdioCodeLocations()) {
             String codeLocationName = bdioCodeLocation.getCodeLocationName();
-            ExternalId externalId = bdioCodeLocation.getDetectCodeLocation().getExternalId();
+            ExternalId detectCodeLocationExternalId = bdioCodeLocation.getDetectCodeLocation().getExternalId();
             DependencyGraph dependencyGraph = bdioCodeLocation.getDetectCodeLocation().getDependencyGraph();
-            ProjectDependencyGraph projectDependencyGraph = new ProjectDependencyGraph(new ProjectDependency(externalId));
+
+            ProjectDependencyGraph projectDependencyGraph = new ProjectDependencyGraph(new ProjectDependency(detectCodeLocationExternalId));
             DependencyGraphUtil.copyRootDependencies(projectDependencyGraph, dependencyGraph);
 
             File bdioOutputFile = new File(outputDirectory, bdioCodeLocation.getBdioName() + ".jsonld");
-            SimpleBdioDocument simpleBdioDocument = simpleBdioFactory.createPopulatedBdioDocument(codeLocationName, projectDependencyGraph);
+
+            BdioNodeFactory bdioNodeFactory = new BdioNodeFactory(new BdioPropertyHelper());
+            BdioProject bdioProject = bdioNodeFactory.createProject(
+                projectNameVersion.getName(),
+                projectNameVersion.getVersion(),
+                detectCodeLocationExternalId.createBdioId(),
+                detectCodeLocationExternalId
+            );
+
+            BdioBillOfMaterials billOfMaterials = bdioNodeFactory.createBillOfMaterials(codeLocationName, bdioProject.name, bdioProject.version);
+            SimpleBdioDocument simpleBdioDocument = new SimpleBdioDocument();
+            simpleBdioDocument.setBillOfMaterials(billOfMaterials);
+            simpleBdioDocument.setProject(bdioProject);
+
+            simpleBdioFactory.populateComponents(simpleBdioDocument, projectDependencyGraph);
+            // TODO: In 8.0.0
+            //  SimpleBdioDocument simpleBdioDocument = simpleBdioFactory.createPopulatedBdioDocument(codeLocationName, projectDependencyGraph);
 
             detectBdioWriter.writeBdioFile(bdioOutputFile, simpleBdioDocument);
             uploadTargets.add(UploadTarget.createDefault(projectNameVersion, codeLocationName, bdioOutputFile));