+++ src/test/java/com/synopsys/integration/detect/battery/util/DetectJar.java
@@ -34,7 +34,13 @@ public class DetectJar {
         File[] libChildren = libs.listFiles();
 
         Assertions.assertNotNull(libChildren, "Could not locate at least one child of libs, does libs exist, was the project built?");
-        Assertions.assertEquals(1, libChildren.length, "Found an unexpected number of jars, only expecting ONE, aka THE JAR expected to run.");
+        // Be sure to clean the project if testing multiple SNAPSHOT versions locally
+        Assertions.assertEquals(
+            1, libChildren.length,
+            "Found an unexpected number of jars, only expecting ONE, aka THE JAR expected to run."
+                + System.lineSeparator() + "Found files: " + System.lineSeparator()
+                + StringUtils.joinWith(System.lineSeparator(), (Object[]) libChildren)
+        );
         File detectJarFile = libChildren[0];
 
         Assertions.assertNotNull(detectJarFile, "Could not find a detect jar!");