+++ src/test/java/com/synopsys/integration/detect/boot/ProductDeciderTest.java
@@ -1,17 +1,14 @@
 package com.synopsys.integration.detect.boot;
 
-import java.nio.file.Path;
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 
-import com.synopsys.integration.common.util.Bds;
 import com.synopsys.integration.detect.configuration.connection.BlackDuckConnectionDetails;
 import com.synopsys.integration.detect.configuration.enumeration.BlackduckScanMode;
 import com.synopsys.integration.detect.lifecycle.boot.decision.BlackDuckDecision;
 import com.synopsys.integration.detect.lifecycle.boot.decision.ProductDecider;
-import com.synopsys.integration.detect.tool.signaturescanner.BlackDuckSignatureScannerOptions;
 import com.synopsys.integration.detect.workflow.bdio.BdioOptions;
 
 class ProductDeciderTest {
@@ -20,199 +17,180 @@ class ProductDeciderTest {
     @Test
     public void shouldRunBlackDuckOfflineWhenOverride() {
         BlackDuckConnectionDetails blackDuckConnectionDetails = blackDuckConnectionDetails(true, null);
-        BlackDuckSignatureScannerOptions blackDuckSignatureScannerOptions = blackDuckSignatureScannerOptions(null, null);
         BlackDuckDecision productDecision = new ProductDecider().decideBlackDuck(
             blackDuckConnectionDetails,
-            blackDuckSignatureScannerOptions,
             BlackduckScanMode.INTELLIGENT,
-            createBdioOptions(false, true)
+            createBdioOptions(false)
         );
 
-        Assertions.assertTrue(productDecision.shouldRun());
-        Assertions.assertTrue(productDecision.isOffline());
+        assertTrue(productDecision.shouldRun());
+        assertTrue(productDecision.isOffline());
     }
 
     @Test
     public void shouldRunOfflineEvenWhenUrlProvided() {
         BlackDuckConnectionDetails blackDuckConnectionDetails = blackDuckConnectionDetails(true, "http://example.com");
-        BlackDuckSignatureScannerOptions blackDuckSignatureScannerOptions = blackDuckSignatureScannerOptions(null, null);
         BlackDuckDecision productDecision = new ProductDecider().decideBlackDuck(
             blackDuckConnectionDetails,
-            blackDuckSignatureScannerOptions,
             BlackduckScanMode.INTELLIGENT,
-            createBdioOptions(false, true)
+            createBdioOptions(false)
         );
 
-        Assertions.assertTrue(productDecision.shouldRun());
-        Assertions.assertTrue(productDecision.isOffline());
+        assertTrue(productDecision.shouldRun());
+        assertTrue(productDecision.isOffline());
     }
 
     @Test
     public void shouldRunBlackDuckOfflineWhenInstallUrl() {
         BlackDuckConnectionDetails blackDuckConnectionDetails = blackDuckConnectionDetails(true, null);
-        BlackDuckSignatureScannerOptions blackDuckSignatureScannerOptions = blackDuckSignatureScannerOptions(null, VALID_URL);
         BlackDuckDecision productDecision = new ProductDecider().decideBlackDuck(
             blackDuckConnectionDetails,
-            blackDuckSignatureScannerOptions,
             BlackduckScanMode.INTELLIGENT,
-            createBdioOptions(false, true)
+            createBdioOptions(false)
         );
 
-        Assertions.assertTrue(productDecision.shouldRun());
-        Assertions.assertTrue(productDecision.isOffline());
+        assertTrue(productDecision.shouldRun());
+        assertTrue(productDecision.isOffline());
     }
 
     @Test
     public void shouldRunBlackDuckOfflineWhenInstallPath() {
         BlackDuckConnectionDetails blackDuckConnectionDetails = blackDuckConnectionDetails(true, null);
-        BlackDuckSignatureScannerOptions blackDuckSignatureScannerOptions = blackDuckSignatureScannerOptions(Mockito.mock(Path.class), null);
         BlackDuckDecision productDecision = new ProductDecider().decideBlackDuck(
             blackDuckConnectionDetails,
-            blackDuckSignatureScannerOptions,
             BlackduckScanMode.INTELLIGENT,
-            createBdioOptions(false, true)
+            createBdioOptions(false)
         );
 
-        Assertions.assertTrue(productDecision.shouldRun());
-        Assertions.assertTrue(productDecision.isOffline());
+        assertTrue(productDecision.shouldRun());
+        assertTrue(productDecision.isOffline());
     }
 
     @Test
     public void shouldNotRunBlackDuckOfflineWhenUserProvidedHostUrl() {
         BlackDuckConnectionDetails blackDuckConnectionDetails = blackDuckConnectionDetails(true, null);
-        BlackDuckSignatureScannerOptions blackDuckSignatureScannerOptions = blackDuckSignatureScannerOptions(Mockito.mock(Path.class), VALID_URL);
         BlackDuckDecision productDecision = new ProductDecider().decideBlackDuck(
             blackDuckConnectionDetails,
-            blackDuckSignatureScannerOptions,
             BlackduckScanMode.INTELLIGENT,
-            createBdioOptions(false, true)
+            createBdioOptions(false)
         );
 
-        Assertions.assertTrue(productDecision.shouldRun());
-        Assertions.assertTrue(productDecision.isOffline());
+        assertTrue(productDecision.shouldRun());
+        assertTrue(productDecision.isOffline());
     }
 
     @Test
     public void shouldRunBlackDuckOnline() {
         BlackDuckConnectionDetails blackDuckConnectionDetails = blackDuckConnectionDetails(false, VALID_URL);
-        BlackDuckSignatureScannerOptions blackDuckSignatureScannerOptions = blackDuckSignatureScannerOptions(null, null);
         BlackDuckDecision productDecision = new ProductDecider().decideBlackDuck(
             blackDuckConnectionDetails,
-            blackDuckSignatureScannerOptions,
             BlackduckScanMode.INTELLIGENT,
-            createBdioOptions(true, true)
+            createBdioOptions(true)
         );
 
-        Assertions.assertTrue(productDecision.shouldRun());
-        Assertions.assertFalse(productDecision.isOffline());
+        assertTrue(productDecision.shouldRun());
+        assertFalse(productDecision.isOffline());
     }
 
     @Test
     public void shouldNotRunBlackduckRapidModeAndOffline() {
         BlackDuckConnectionDetails blackDuckConnectionDetails = blackDuckConnectionDetails(true, VALID_URL);
-        BlackDuckSignatureScannerOptions blackDuckSignatureScannerOptions = blackDuckSignatureScannerOptions(null, null);
         BlackDuckDecision productDecision = new ProductDecider().decideBlackDuck(
             blackDuckConnectionDetails,
-            blackDuckSignatureScannerOptions,
             BlackduckScanMode.RAPID,
-            createBdioOptions(false, true)
+            createBdioOptions(false)
         );
 
-        Assertions.assertFalse(productDecision.shouldRun());
+        assertFalse(productDecision.shouldRun());
     }
 
     @Test
     public void shouldNotRunBlackduckRapidModeAndBDIO2Disabled() {
         BlackDuckConnectionDetails blackDuckConnectionDetails = blackDuckConnectionDetails(false, VALID_URL);
-        BlackDuckSignatureScannerOptions blackDuckSignatureScannerOptions = blackDuckSignatureScannerOptions(null, null);
         BlackDuckDecision productDecision = new ProductDecider().decideBlackDuck(
             blackDuckConnectionDetails,
-            blackDuckSignatureScannerOptions,
             BlackduckScanMode.RAPID,
-            createBdioOptions(false, true)
+            createBdioOptions(false)
         );
 
-        Assertions.assertFalse(productDecision.shouldRun());
+        assertFalse(productDecision.shouldRun());
     }
 
     @Test
     public void shouldNotRunBlackduckIntelligentModeAndBDIO2Disabled() {
         BlackDuckConnectionDetails blackDuckConnectionDetails = blackDuckConnectionDetails(false, VALID_URL);
-        BlackDuckSignatureScannerOptions blackDuckSignatureScannerOptions = blackDuckSignatureScannerOptions(null, null);
         BlackDuckDecision productDecision = new ProductDecider().decideBlackDuck(
             blackDuckConnectionDetails,
-            blackDuckSignatureScannerOptions,
             BlackduckScanMode.INTELLIGENT,
-            createBdioOptions(false, false)
+            createBdioOptions(false)
         );
 
-        Assertions.assertFalse(productDecision.shouldRun());
+        assertFalse(productDecision.shouldRun());
     }
 
     @Test
     public void shouldRunBlackduckRapidModeAndBDIO2Enabled() {
         BlackDuckConnectionDetails blackDuckConnectionDetails = blackDuckConnectionDetails(false, VALID_URL);
-        BlackDuckSignatureScannerOptions blackDuckSignatureScannerOptions = blackDuckSignatureScannerOptions(null, null);
         BlackDuckDecision productDecision = new ProductDecider().decideBlackDuck(
             blackDuckConnectionDetails,
-            blackDuckSignatureScannerOptions,
             BlackduckScanMode.RAPID,
-            createBdioOptions(true, true)
+            createBdioOptions(true)
         );
 
-        Assertions.assertTrue(productDecision.shouldRun());
+        assertTrue(productDecision.shouldRun());
+    }
+
+    @Test
+    public void shouldRunBlackduckIntelligentModeOfflineAndBDIO2Disabled() {
+        BlackDuckConnectionDetails blackDuckConnectionDetails = blackDuckConnectionDetails(true, VALID_URL);
+        BlackDuckDecision productDecision = new ProductDecider().decideBlackDuck(
+            blackDuckConnectionDetails,
+            BlackduckScanMode.INTELLIGENT,
+            createBdioOptions(false)
+        );
+
+        assertTrue(productDecision.shouldRun());
+        assertTrue(productDecision.isOffline());
     }
 
     @Test
     public void shouldRunBlackduckIntelligentModeAndBDIO2Enabled() {
         BlackDuckConnectionDetails blackDuckConnectionDetails = blackDuckConnectionDetails(false, VALID_URL);
-        BlackDuckSignatureScannerOptions blackDuckSignatureScannerOptions = blackDuckSignatureScannerOptions(null, null);
         BlackDuckDecision productDecision = new ProductDecider().decideBlackDuck(
             blackDuckConnectionDetails,
-            blackDuckSignatureScannerOptions,
             BlackduckScanMode.INTELLIGENT,
-            createBdioOptions(true, false)
+            createBdioOptions(true)
         );
 
-        Assertions.assertTrue(productDecision.shouldRun());
+        assertTrue(productDecision.shouldRun());
     }
 
     @Test
-    public void shouldRunBlackduckLegacyEnabledAndIntelligentModeAndBDIO2Disabled() {
+    public void shouldNotRunBlackduckOnlineIntelligentModeAndBDIO2Disabled() {
         BlackDuckConnectionDetails blackDuckConnectionDetails = blackDuckConnectionDetails(false, VALID_URL);
-        BlackDuckSignatureScannerOptions blackDuckSignatureScannerOptions = blackDuckSignatureScannerOptions(null, null);
         BlackDuckDecision productDecision = new ProductDecider().decideBlackDuck(
             blackDuckConnectionDetails,
-            blackDuckSignatureScannerOptions,
             BlackduckScanMode.INTELLIGENT,
-            createBdioOptions(false, true)
+            createBdioOptions(false)
         );
 
-        Assertions.assertTrue(productDecision.shouldRun());
+        assertFalse(productDecision.shouldRun());
     }
 
     @Test
     public void shouldNotRunBlackduckURLMissing() {
         BlackDuckConnectionDetails blackDuckConnectionDetails = blackDuckConnectionDetails(false, null);
-        BlackDuckSignatureScannerOptions blackDuckSignatureScannerOptions = blackDuckSignatureScannerOptions(null, null);
         BlackDuckDecision productDecision = new ProductDecider().decideBlackDuck(
             blackDuckConnectionDetails,
-            blackDuckSignatureScannerOptions,
             BlackduckScanMode.RAPID,
-            createBdioOptions(true, true)
+            createBdioOptions(true)
         );
 
-        Assertions.assertFalse(productDecision.shouldRun());
+        assertFalse(productDecision.shouldRun());
     }
 
-    private BdioOptions createBdioOptions(boolean useBdio2, boolean enableLegacyUpload) {
-        return new BdioOptions(useBdio2, null, null, enableLegacyUpload);
-    }
-
-    private BlackDuckSignatureScannerOptions blackDuckSignatureScannerOptions(Path offlineScannerInstallPath, String userProvidedScannerInstallUrl) {
-        return new BlackDuckSignatureScannerOptions(Bds.listOf(), Bds.listOf(), null, 1024, 1, false, null, false, null, null, null, 1, null, false, false, true,
-            true
-        );
+    private BdioOptions createBdioOptions(boolean useBdio2) {
+        return new BdioOptions(useBdio2, null, null);
     }
 
     private BlackDuckConnectionDetails blackDuckConnectionDetails(boolean offline, String blackduckUrl) {